updated comments from Micheal
applied all suggestions:

In dce_data.py : please could you update the docstrings to describe the tests/data?
Suggest the function names in the above module should reflect the data sources + models
Some of the modules contain many commented lines - can we remove?
One of the test function in the OGJ module is named LEK
It's unclear why various test cases are excluded for each contribution
In the CSV data files, I suggest putting the parameter columns next to the labels column, to make it more readable for the user